backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
chromadb==1.3.0
python-dotenv==1.0.1
pydantic==2.9.2
openai==1.52.2
psycopg2-binary==2.9.9

# MCP Server
fastmcp>=0.9.0

# Future dependencies for web scraping (to be added in Phase 5):
# crawl4ai==0.6.2
# requests>=2.31.0
# PyGithub>=2.1.0
# gitpython>=3.1.40
